Edited some potential errors
John & Tarin stepped through Schedule.h looking for potential bugs and errors. So far this has just been a logical stepping through the code looking for potential bugs. Actual fixes and updates along with testing that the Schedule class works correctly is a work in progress (WIP) but will be updated later.

Likewise, these changes were pushed to a new branch - "testing" in the event that any current changes cause a conflict with any other changes and to ensure other teammates do not accidentally overwrite the changes that were made (if all changes are just pushed to the master branch).

Once working correctly the branches will be merged.