Update Tarin test code.txt
This is a basic prototype. It assumes that the Node known as head exists. It then checks if head is null or not. If null it tells the user that there are no events. If head is not null, and thus there are events, it prints two lines. One contains the event/alarm label, the other contains the date in integer form

I want to print the date out later as an actual date, but let's save our work first